+++ packages/website/pages/_app.js
@@ -15,8 +15,7 @@ import Footer from '../components/footer/footer.js';
  */
 const App = ({ Component, pageProps }) => {
   const { pathname } = useRouter();
-  const productRoutes = ['/login', '/account', '/tokens', '/callback'];
-  // const marketingRoutes = ['/', '/pricing', '/about', '/faq', '/terms'];
+  const productRoutes = ['/login', '/account', '/account/payment', '/tokens', '/callback'];
   const productApp = productRoutes.includes(pathname);
   const pageClass = pathname.includes('docs') ? 'docs-site' : productApp ? 'product-app' : 'marketing-site';