+++ packages/website/pages/tokens/index.js
@@ -42,7 +42,7 @@ const Tokens = () => {
           href={content.ui.return.url}
           variant={content.ui.return.theme}
         >
-          <Link href={content.ui.return.url}>{content.ui.return.text}</Link>
+          {content.ui.return.text}
         </Button>
         <div dangerouslySetInnerHTML={{ __html: content.ui.test_token }} className="testing-cta-container"></div>
       </div>